--- include/userInputParameters_template_instantiations.h
@@ -1,7 +1,0 @@
-#ifndef INCLUDE_USERINPUTPARAMETERS_TEMPLATE_INSTANTIATIONS_H_
-#define INCLUDE_USERINPUTPARAMETERS_TEMPLATE_INSTANTIATIONS_H_
-
-template class userInputParameters<2>;
-template class userInputParameters<3>;
-
-#endif
