+++ src/variableAttributeLoader/variableAttributeLoader.cc
@@ -3,6 +3,7 @@
 #include <deal.II/base/exceptions.h>
 #include <deal.II/base/utilities.h>
 
+// NOLINTBEGIN(cppcoreguidelines-prefer-member-initializer)
 variableAttributeLoader::variableAttributeLoader()
 {
   relevant_attributes = &attributes;
@@ -44,44 +45,47 @@ variableAttributeLoader::variableAttributeLoader()
     }
 }
 
+// NOLINTEND(cppcoreguidelines-prefer-member-initializer)
+
 // Methods to set the various variable attributes
 void
 variableAttributeLoader::set_variable_name(const unsigned int &index,
-                                           const std::string  &name)
+                                           const std::string  &name) const
 {
   (*relevant_attributes)[index].name = name;
 }
 
 void
 variableAttributeLoader::set_variable_type(const unsigned int &index,
-                                           const fieldType    &var_type)
+                                           const fieldType    &var_type) const
 {
   (*relevant_attributes)[index].var_type = var_type;
 }
 
 void
 variableAttributeLoader::set_variable_equation_type(const unsigned int &index,
-                                                    const PDEType      &var_eq_type)
+                                                    const PDEType      &var_eq_type) const
 {
   (*relevant_attributes)[index].eq_type = var_eq_type;
 }
 
 void
 variableAttributeLoader::set_need_value_nucleation(const unsigned int &index,
-                                                   const bool         &flag)
+                                                   const bool         &flag) const
 {
   (*relevant_attributes)[index].need_value_nucleation = flag;
 }
 
 void
 variableAttributeLoader::set_allowed_to_nucleate(const unsigned int &index,
-                                                 const bool         &flag)
+                                                 const bool         &flag) const
 {
   (*relevant_attributes)[index].nucleating_variable = flag;
 }
 
 void
-variableAttributeLoader::set_output_integral(const unsigned int &index, const bool &flag)
+variableAttributeLoader::set_output_integral(const unsigned int &index,
+                                             const bool         &flag) const
 {
   (*relevant_attributes)[index].output_integral = flag;
   (*relevant_attributes)[index].calc_integral   = flag;
@@ -129,8 +133,9 @@ variableAttributeLoader::set_dependencies_gradient_term_RHS(
 }
 
 void
-variableAttributeLoader::set_dependencies_value_term_LHS(const unsigned int &index,
-                                                         const std::string  &dependencies)
+variableAttributeLoader::set_dependencies_value_term_LHS(
+  const unsigned int &index,
+  const std::string  &dependencies) const
 {
   std::vector<std::string> dependencies_set =
     dealii::Utilities::split_string_list(strip_whitespace(dependencies));
@@ -141,12 +146,111 @@ variableAttributeLoader::set_dependencies_value_term_LHS(const unsigned int &ind
 void
 variableAttributeLoader::set_dependencies_gradient_term_LHS(
   const unsigned int &index,
-  const std::string  &dependencies)
+  const std::string  &dependencies) const
 {
   std::vector<std::string> dependencies_set =
     dealii::Utilities::split_string_list(strip_whitespace(dependencies));
   (*relevant_attributes)[index].dependencies_gradient_LHS =
     std::set<std::string>(dependencies_set.begin(), dependencies_set.end());
+}
+
+void
+variableAttributeLoader::validate_variable_name(
+  const std::string           &name,
+  const std::set<std::string> &forbidden_names,
+  const std::string           &context,
+  unsigned int                 index)
+{
+  Assert(!name.empty(),
+         dealii::ExcMessage("PRISMS-PF Error: " + context +
+                            " Variable names must not be empty.\nProblem index: " +
+                            std::to_string(index)));
+
+  for (const std::string &forbidden_name : forbidden_names)
+    {
+      std::string error_message = "PRISMS-PF Error: " + context +
+                                  " Variable names must not contain \"grad()\", "
+                                  "\"hess()\", \"change()\".\nProblem index: ";
+      error_message.append(std::to_string(index).append(", Variable name: " + name));
+
+      Assert(name.find(forbidden_name) == std::string::npos,
+             dealii::ExcMessage(error_message));
+    }
+}
+
+void
+variableAttributeLoader::populate_dependencies(
+  const std::set<std::pair<std::string, std::string>> &reg_delimiters,
+  const std::string                                   &variable_name,
+  unsigned int                                         index,
+  std::set<std::string>                               &reg_possible_deps,
+  std::map<uint, std::set<std::string>>               &change_possible_deps)
+{
+  for (const auto &delims : reg_delimiters)
+    {
+      reg_possible_deps.insert(delims.first + variable_name + delims.second);
+      change_possible_deps[index].insert(delims.first + "change(" + variable_name + ")" +
+                                         delims.second);
+    }
+}
+
+void
+variableAttributeLoader::validate_dependencies(
+  const std::set<std::string>                 &dependencies,
+  const std::string                           &context,
+  unsigned int                                 index,
+  const std::string                           &variable_name,
+  const std::set<std::string>                 &reg_possible_deps,
+  const std::map<uint, std::set<std::string>> &change_possible_deps)
+{
+  for (const std::string &dependency : dependencies)
+    {
+      std::string error_message =
+        "PRISMS-PF Error: Invalid " + context + " dependency.\nProblem index: ";
+      error_message.append(
+        std::to_string(index).append(", Variable name: " + variable_name));
+      error_message.append(
+        ", Invalid dependency name: " + dependency +
+        "\n(HINT: Dependencies that contain \"change(variable_A)\" can only be "
+        "used for the field \"variable_A\")\n");
+
+      Assert(reg_possible_deps.find(dependency) != reg_possible_deps.end() ||
+               change_possible_deps.at(index).find(dependency) !=
+                 change_possible_deps.at(index).end(),
+             dealii::ExcMessage(error_message));
+    }
+}
+
+void
+variableAttributeLoader::variableAttributeLoader::validate_postprocess_variable(
+  const std::string           &name,
+  const std::set<std::string> &name_list,
+  const std::set<std::string> &reg_possible_deps,
+  const variableAttributes    &pp_variable,
+  unsigned int                 index)
+{
+  Assert(name_list.find(name) == name_list.end(),
+         dealii::ExcMessage("PRISMS-PF Error: Postprocess variable names must be "
+                            "different from solution variable names.\nProblem index: " +
+                            std::to_string(index) + ", Variable name: " + name + "\n"));
+  for (const auto &PP_dep : pp_variable.dependencies_PP)
+    {
+      std::string error_message =
+        "PRISMS-PF Error: Invalid postprocessing RHS dependency.\nProblem index: ";
+      error_message.append(std::to_string(index).append(", Variable name: " + name));
+      error_message.append(", Invalid dependency name: " + PP_dep + "\n");
+
+      Assert(reg_possible_deps.find(PP_dep) != reg_possible_deps.end(),
+             dealii::ExcMessage(error_message));
+    }
+  Assert(pp_variable.dependencies_LHS.empty(),
+         dealii::ExcMessage("PRISMS-PF Error: Postprocess variables can only have RHS "
+                            "dependencies.\nProblem index: " +
+                            std::to_string(index) + ", Variable name: " + name + "\n"));
+  Assert(!pp_variable.nucleating_variable && !pp_variable.need_value_nucleation,
+         dealii::ExcMessage("PRISMS-PF Error: Postprocess variables cannot be used for "
+                            "nucleation.\nProblem index: " +
+                            std::to_string(index) + ", Variable name: " + name + "\n"));
 }
 
 void
@@ -161,108 +265,58 @@ variableAttributeLoader::validate_attributes()
     {"hess(", ")"}
   };
   const std::set<std::string>           forbidden_names = {"grad(", "hess(", "change("};
-  std::set<std::string>                 name_list, pp_name_list, reg_possible_deps;
+  std::set<std::string>                 name_list;
+  std::set<std::string>                 pp_name_list;
+  std::set<std::string>                 reg_possible_deps;
   std::map<uint, std::set<std::string>> change_possible_deps;
-  // Populate name_list, pp_name_list, reg_possible_deps, change_possible_deps
-  // Check that variable names are mostly well-formed
+
+  // Populate the expected variable dependencies and check that variable names are mostly
+  // well-formed.
   for (const auto &[index, variable] : attributes)
     {
       name_list.insert(variable.name);
-      for (const auto &delims : reg_delimiters)
-        {
-          reg_possible_deps.insert(delims.first + variable.name + delims.second);
-          change_possible_deps[index].insert(delims.first + "change(" + variable.name +
-                                             ")" + delims.second);
-        }
-      Assert(!variable.name.empty(),
-             dealii::ExcMessage(
-               "PRISMS-PF Error: Variable names must not be empty.\nProblem index: " +
-               std::to_string(index) + "\n"));
-      for ([[maybe_unused]] const std::string &forbidden_name : forbidden_names)
-        {
-          Assert(variable.name.find(forbidden_name) == std::string::npos,
-                 dealii::ExcMessage(
-                   "PRISMS-PF Error: Variable names must not contain \"grad()\", "
-                   "\"hess()\", \"change()\".\nProblem index: " +
-                   std::to_string(index) + ", Variable name: " + variable.name + "\n"));
-        }
+
+      populate_dependencies(reg_delimiters,
+                            variable.name,
+                            index,
+                            reg_possible_deps,
+                            change_possible_deps);
+
+      validate_variable_name(variable.name, forbidden_names, "", index);
     }
+
+  // Validate the postprocessed variables
   for (const auto &[pp_index, pp_variable] : pp_attributes)
     {
       pp_name_list.insert(pp_variable.name);
-      Assert(!pp_variable.name.empty(),
-             dealii::ExcMessage(
-               "PRISMS-PF Error: (postprocess) Variable names must not be "
-               "empty.\nProblem index: " +
-               std::to_string(pp_index) + "\n"));
-      for ([[maybe_unused]] const std::string &forbidden_name : forbidden_names)
-        {
-          Assert(pp_variable.name.find(forbidden_name) == std::string::npos,
-                 dealii::ExcMessage("PRISMS-PF Error: (postprocess) Variable names must "
-                                    "not contain \"grad()\", "
-                                    "\"hess()\", \"change()\".\nProblem index: " +
-                                    std::to_string(pp_index) +
-                                    ", Variable name: " + pp_variable.name + "\n"));
-        }
+
+      validate_variable_name(pp_variable.name,
+                             forbidden_names,
+                             "(postprocess)",
+                             pp_index);
+
+      validate_postprocess_variable(pp_variable.name,
+                                    name_list,
+                                    reg_possible_deps,
+                                    pp_variable,
+                                    pp_index);
     }
-  // Check Dependencies & PP
+  // Check dependencies
   for (const auto &[index, variable] : attributes)
     {
-      for ([[maybe_unused]] const std::string &RHS_dep : variable.dependencies_RHS)
-        {
-          Assert(
-            reg_possible_deps.find(RHS_dep) != reg_possible_deps.end() ||
-              change_possible_deps.at(index).find(RHS_dep) !=
-                change_possible_deps.at(index).end(),
-            dealii::ExcMessage(
-              "PRISMS-PF Error: Invalid RHS dependency.\nProblem index: " +
-              std::to_string(index) + ", Variable name: " + variable.name +
-              ", Invalid dependency name: " + RHS_dep +
-              "\n(HINT: Dependencies that contain \"change(variable_A)\" can only be "
-              "used for the "
-              "field \"variable_A\")\n"));
-        }
-      for ([[maybe_unused]] const std::string &LHS_dep : variable.dependencies_LHS)
-        {
-          Assert(
-            reg_possible_deps.find(LHS_dep) != reg_possible_deps.end() ||
-              change_possible_deps.at(index).find(LHS_dep) !=
-                change_possible_deps.at(index).end(),
-            dealii::ExcMessage(
-              "PRISMS-PF Error: Invalid LHS dependency.\nProblem index: " +
-              std::to_string(index) + ", Variable name: " + variable.name +
-              ", Invalid dependency name: " + LHS_dep +
-              "\n(HINT: Dependencies that contain \"change(variable_A)\" can only be "
-              "used for the "
-              "field \"variable_A\")\n"));
-        }
-    }
-  for (const auto &[pp_index, pp_variable] : pp_attributes)
-    {
-      Assert(name_list.find(pp_variable.name) == name_list.end(),
-             dealii::ExcMessage(
-               "PRISMS-PF Error: Postprocess variable names must be named differently "
-               "than solution variable names.\nProblem postprocessing index: " +
-               std::to_string(pp_index) + ", Variable name: " + pp_variable.name + "\n"));
-      for ([[maybe_unused]] const std::string &PP_dep : pp_variable.dependencies_PP)
-        {
-          Assert(reg_possible_deps.find(PP_dep) != reg_possible_deps.end(),
-                 dealii::ExcMessage(
-                   "PRISMS-PF Error: Invalid postprocessing RHS dependency.\nProblem "
-                   "postprocessing index: " +
-                   std::to_string(pp_index) + ", Variable name: " + pp_variable.name +
-                   ", Invalid dependency name: " + PP_dep + "\n"));
-        }
-      Assert(pp_variable.dependencies_LHS.empty(),
-             dealii::ExcMessage(
-               "PRISMS-PF Error: Warning: Postprocess variables can only have RHS "
-               "dependencies.\nProblem postprocessing index: " +
-               std::to_string(pp_index) + ", Variable name: " + pp_variable.name + "\n"));
-      Assert(!pp_variable.nucleating_variable && !pp_variable.need_value_nucleation,
-             dealii::ExcMessage(
-               "PRISMS-PF Error: Warning: Postprocess variables cannot "
-               "be used for nucleation.\nProblem postprocessing index: " +
-               std::to_string(pp_index) + ", Variable name: " + pp_variable.name + "\n"));
+      validate_dependencies(variable.dependencies_RHS,
+                            "RHS",
+                            index,
+                            variable.name,
+                            reg_possible_deps,
+                            change_possible_deps);
+
+      validate_dependencies(variable.dependencies_LHS,
+                            "LHS",
+                            index,
+                            variable.name,
+                            reg_possible_deps,
+                            change_possible_deps);
     }
 }
