+++ tests/unit_tests/unitTest.h
@@ -59,10 +59,6 @@ computeStress(const dealii::Table<2, double>       &CIJ,
 #include "../../src/matrixfree/utilities.cc"
 #include "../../src/models/mechanics/computeStress.h"
 #include "../../src/parallelNucleationList/parallelNucleationList.cc"
-#include "../../src/userInputParameters/loadVariableAttributes.cc"
-#include "../../src/userInputParameters/load_BC_list.cc"
-#include "../../src/userInputParameters/load_user_constants.cc"
-#include "../../src/userInputParameters/setTimeStepList.cc"
 #include "../../src/variableAttributeLoader/variableAttributeLoader.cc"
 #include "../../src/variableContainer/variableContainer.cc"
